requirements-dev.txt
-r requirements.txt
pytest>=8.3.0
moto[dynamodb]>=5.0.0
responses>=0.25.0
python-dotenv>=1.0.0
ruff>=0.8.0
mypy>=1.13.0
boto3-stubs[dynamodb]
